[imdb] Fix font escape sequences
If the IMDb text contains any of the strings "\c", "\n", "\r" or
"\t", they are interpreted as escape sequences by
eTextPara::renderString().

Unfortunately, renderString() does not provide a clean way to escape
the "\" character, so the "\" that is to appear in the text is
separated from the alpha character by a <CR>, which is a no-op in
eTextPara::renderString().

So, for example, "\c" in the IMDb text is transformed into "\<CR>c",
which then displays as "\c" on the screen.

There really should be a clean way to escape the escape character
in renderString().

The change is prompted by real data in IMDB. In the Trivia section
the movie "The Core" (2003) a commenter uses "\" where "/" would
normally be used: "In the 1950s\60s, during the design\construction
of the SR-71 Blackbird ...". The "\c..." is rendered as a colour
change.