AutoTimer: init_py: builtins.RecursionError: maximum recursion depth exceeded
raise self.value.with_traceback(self.tb)
in <lambda>
in any case please use automatic timers, you have to observe the behavior.
There is a very easy way now to do it, when you are watching an event on TV, press the yellow button and add a self-timer.
then look for it with the green button in Auto Timers.